Fix replace conflict strategy
After initially failing to get this feature working, I found that the strategy class wasn't in the hash of strategies.

The other thing this PR addresses is removing the client middleware is loaded on the server.
The client middleware checks if a lock is in place and enqueues in its absense. Client middleware is only hit when
the server enqueues to itself which happens when sidekiq takes scheduled items and enqueues them on the work queue.
I was trying to use the `:until_executing` lock along with `:replace` strategy, but the server would just take the job
and drop in `Client::Middleware#call` because the lock only gets released during the server middleware call.

All ears if the client middleware on the server is necessary for whatever other reason.

All the tests pass without it (bar the test which just checks the presence of it, which I fixed).