Log job silently complete
Borrows heavily from https://github.com/mhenrixon/sidekiq-unique-jobs/pull/320
with the objective to get it shipped.

The problem, as described, is that on a surface a job successfully
completes where in fact it wasn't executed at all, due to a lock.
This commit adds some clarity in the form of a log message explaining
that the job did complete silently.