Fix return value of `base_lock`
When using the `on_conflict` `replace` strategy `base_lock` has to
return the return value of the conflict strategy. The replace strategy
removes the previous job from the scheduled/rety set or the queue.
It is then re-queued inside `lock_failed`/`call_strategy`, thus we have
to return the return value of which came from `lock` inside the block of
` client_strategy.call`.