Fixing reschedule when using a non default queue (#679)
* `on_conflict { server: :reschedule }` always uses default queue.

I noticed, that when I use the `on_conflict` strategy `:reschedule` with a job, that has a
`until_and_while_executing` lock and is configured to run on a specific
queue,(e.g. `no_default`). `:reschedule` tries to enqueue it into the
`default` queue.
The commit lets `:reschedule` honour the queue.

* Modify specs to honour queue.

* Listen to Rubocop, wise he is.