Add lock to item when lock_type is pulled from Job class sidekiq options
For jobs that are queued without the lock options, to a Job class that has `lock` option set in the `sidekiq_options` will still go through the SidekiqUniqueJobs server middleware, but won't have `lock` set on the item the entire time.
This doesn't seem to be affecting the locking or unlocking of jobs, however it is affecting the logging_context because the Middleware LogContext expects the item[LOCK] to have a value.